refactor(gateway): extract GatewayNotificationsMixin from run.py (slice 28 of #54962)
